+++ src/main/java/com/novell/ldapchai/impl/directoryServer389/entry/DirectoryServer389VendorFactory.java
@@ -44,6 +44,17 @@ public class DirectoryServer389VendorFactory implements VendorFactory
 
     private static final ErrorMap ERROR_MAP = new EdirErrorMap();
 
+    private static final DirectoryServer389VendorFactory SINGLETON = new DirectoryServer389VendorFactory();
+
+    public static DirectoryServer389VendorFactory getInstance()
+    {
+        return SINGLETON;
+    }
+
+    private DirectoryServer389VendorFactory()
+    {
+    }
+
     @Override
     public ChaiUser newChaiUser( final String entryDN, final ChaiProvider provider )
     {